Added a "Settings" button to the main activity.  Whne this activity stops becasue the settings button is pressed it does not disconnect from the service, whcih continues to run, because it is needed by the settigns screen.
Changed the setting screen to use the new format for sending updates to the color blob finder on the acr.  This is then transparently forwarded directly to the Camera.
Need to ajust the other sent commands to fit in with this approach - or to design a proper protocol from Andrid to the Bluno.