Added more to the send/receve messages between the Bluno and Android =============================================================== Added a background image (trivial) and also now send a start character (!) and other initialisation messages when a connection is made.  This includes sending the preference values of the lane sensor colour.
This allows the initialisation of the car to pause before a connection is made so that initialisation messages are seen on the Android device.

Later on the (!) cahracter is just ignored by the car, but a new connection puts it into a known state.

Also process [EInnn] messages as distance messages from sensor I - distance is nnn: these are displayed on a little picture of a yellow car.